pyproject.toml
[tool.poetry]
name = "my-money-smart"
version = "0.1.0"
description = ""
authors = ["vellimir <kkazancov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.8"
djangorestframework-simplejwt = "^5.4.0"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
